Init vue router framework.
