+++ web/src/main.ts
@@ -1,10 +1,11 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
+import router from '@/router';
 import App from '@/App.vue';
 
 const app = createApp(App);
 
-const pinia = createPinia();
-app.use(pinia);
+app.use(createPinia());
+app.use(router);
 
 app.mount('#app');
