+++ web/src/views/LogView.vue
@@ -1,0 +1,3 @@
+<template>
+    <div></div>
+</template>
