+++ web/src/views/PreviewView.vue
@@ -1,0 +1,3 @@
+<template>
+    <div></div>
+</template>
